refactor-> use useGetData for hospital
