+++ src/hooks/generic/useGetData.ts
@@ -1,4 +1,4 @@
-import { useQuery } from "react-query";
+import { useQuery } from "@tanstack/react-query";
 import apiClient from "../../services/axiosClient";
 
 export interface FetchResponse<T> {
@@ -11,26 +11,41 @@ const useGetData = <T>(
   queryKey: string,
   query: string = ""
 ) => {
-  const fetchData = async (): Promise<T | FetchResponse<T>> => {
+  const fetchData = async (): Promise<T> => {
     const authToken = localStorage.getItem("token");
     const headers: Record<string, string> = authToken
       ? { Authorization: `Bearer ${authToken}` }
       : {};
 
-    const response = await apiClient.get<T | FetchResponse<T>>(
-      `${endpoint}?search=${query}`,
-      {
-        headers,
-      }
-    );
+    const response = await apiClient.get<T>(`${endpoint}?search=${query}`, {
+      headers,
+    });
 
     return response.data;
   };
 
-  return useQuery([queryKey, endpoint, query], fetchData, {
-    staleTime: 60000,
-    retry: 2,
-  });
+  const { data, isLoading, error, refetch } = useQuery(
+    [queryKey, endpoint, query],
+    fetchData,
+    {
+      staleTime: 60000,
+      retry: 2,
+    }
+  );
+
+  const getErrorMessage = (error: unknown): string => {
+    if (error instanceof Error) {
+      return error.message;
+    }
+    return "An unknown error occurred";
+  };
+
+  return {
+    data,
+    isLoading,
+    error: error ? getErrorMessage(error) : null,
+    refetch,
+  };
 };
 
 export default useGetData;
